# lib/rand: Optimizations